pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-corptools-dashboard"
description = "Dashboard Addon for Corp Tools"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Geuthur", email = "devgeuthur@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.12,<5",
    "allianceauth-corp-tools",
]
optional-dependencies.tests-allianceauth-latest = [
    "coverage",
    "factory-boy",
]
urls.Changelog = "https://github.com/Geuthur/aa-corptools-dashboard/blob/master/CHANGELOG.md"
urls.Documentation = "https://github.com/Geuthur/aa-corptools-dashboard/blob/master/README.md"
urls.Donations = "https://ko-fi.com/Geuthur"
urls.Homepage = "https://github.com/Geuthur/aa-corptools-dashboard"
urls.Source = "https://github.com/Geuthur/aa-corptools-dashboard"
urls.Tracker = "https://github.com/Geuthur/aa-corptools-dashboard/issues"

[tool.hatch.build]
include = [
    "/ct_dashboard",
]

[tool.hatch.version]
path = "ct_dashboard/__init__.py"
